requirements.txt
streamlit
pandas
plotly.express
bcrypt
st_pages